services/web/requirements.txt
alembic==1.10.4
Flask==2.3.2
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==2.0.11
pydub==0.25.1
base32-lib==1.0.2
psycopg2-binary==2.9.6
pytest==7.3.1